+++ src/components/Modals/HotkeysModal/HotkeysModal.tsx
@@ -1,29 +1,10 @@
 import { Modals, useModalStore } from "@/src/hooks/stores/useModalStore";
-import useKeyPress, { hotkeys } from "@/src/hooks/useKeyPress";
+import useKeyPress from "@/src/hooks/useKeyPress";
+import { hotkeyDescriptions, hotkeys } from "@/src/services/hotkeys/hotkeys";
 
-import Hotkey from "./partials/Hotkey";
+import { Hotkey } from "./partials/HotkeyDescription";
 import Button from "../../ui/Button";
 import Modal from "../../ui/Modal";
-
-export type HotkeyType = {
-  k: string | string[];
-  description: string;
-};
-
-const items: HotkeyType[] = [
-  {
-    description: "Toggle sidebar visibility",
-    k: hotkeys.backslash,
-  },
-  {
-    description: "Toggle the menu you're currently looking at",
-    k: hotkeys.questionMark,
-  },
-  {
-    description: "Toggle the settings menu",
-    k: hotkeys.slash,
-  },
-];
 
 const HotkeysModal = () => {
   const activeModal = useModalStore((state) => state.activeModal);
@@ -38,15 +19,17 @@ const HotkeysModal = () => {
       isVisible={activeModal === Modals.hotkey}
       onClose={() => setActiveModal(null)}
       title="Keyboard Shortcuts"
-      subtitle="Speed up your UpTab game with shortcuts. So much time for activities."
+      subtitle="Speed up your UpTab game with hotkeys. So much time for activities."
     >
       <div className="flex flex-col">
-        {items.map((item) => (
-          <Hotkey
-            key={String(item.k)}
-            k={item.k}
-            description={item.description}
-          />
+        {Object.entries(hotkeyDescriptions).map(([key, item]) => (
+          <div
+            key={key}
+            className="mt-2 flex items-center justify-between border-t pt-2 first:mt-0 first:border-t-0"
+          >
+            <p className="text-gray-900">{item.description}</p>
+            <Hotkey label={item.keyDisplay} size="sm" />
+          </div>
         ))}
         <div className="mt-8 flex justify-end gap-2">
           <Button