+++ src/components/Modals/ShortlinksModal/ShortlinksModal.tsx
@@ -1,21 +1,30 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useEffect, useMemo } from "react";
-import { SubmitHandler, useForm } from "react-hook-form";
+import { SubmitHandler, useForm, useWatch } from "react-hook-form";
 import { v4 as uuidv4 } from "uuid";
 
+import { SHORTLINK_QUERY_KEY } from "@/src/hooks/queries/useShortlinks";
 import { Modals, useModalStore } from "@/src/hooks/stores/useModalStore";
 import { useShortlinkStore } from "@/src/hooks/stores/useShortlinkStore";
-import { addShortlink, deleteShortlink } from "@/src/services/shortlinks";
+import {
+  addShortlink,
+  deleteShortlink,
+  editShortlink,
+  Shortlink as ShortlinkType,
+} from "@/src/services/shortlinks";
+import { getFavicon } from "@/src/utils/getFavicon";
 import { normalizeUrl } from "@/src/utils/normalizeUrl";
 
 import FormField from "./partials/FormField";
+import Shortlink from "../../Shortlinks/partials/Shortlink";
 import Button, { buttonVariants } from "../../ui/Button";
 import Modal from "../../ui/Modal";
 
 export interface FormInputs {
-  URL: string;
-  Title: string;
-  Subtitle: string;
+  url: string;
+  title: string;
+  subtitle: string;
+  icon: string;
 }
 
 const ShortlinksModal = () => {
@@ -24,14 +33,21 @@ const ShortlinksModal = () => {
   const addMutation = useMutation({
     mutationFn: addShortlink,
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["shortlinks"] });
+      queryClient.invalidateQueries({ queryKey: SHORTLINK_QUERY_KEY });
+    },
+  });
+
+  const editMutation = useMutation({
+    mutationFn: editShortlink,
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: SHORTLINK_QUERY_KEY });
     },
   });
 
   const deleteMutation = useMutation({
     mutationFn: deleteShortlink,
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["shortlinks"] });
+      queryClient.invalidateQueries({ queryKey: SHORTLINK_QUERY_KEY });
     },
   });
 
@@ -42,27 +58,53 @@ const ShortlinksModal = () => {
 
   const defaultValues = useMemo(
     () => ({
-      Subtitle: selectedShortlink?.subtitle ?? "",
-      Title: selectedShortlink?.title ?? "",
-      URL: selectedShortlink?.url ?? "",
+      subtitle: selectedShortlink?.subtitle ?? "",
+      title: selectedShortlink?.title ?? "",
+      url: selectedShortlink?.url ?? "",
+      icon: selectedShortlink?.icon ?? "",
     }),
-    [selectedShortlink]
+    [selectedShortlink],
   );
 
-  const { register, handleSubmit, reset } = useForm<FormInputs>({
+  const {
+    register,
+    handleSubmit,
+    reset,
+    control,
+    formState: { errors },
+  } = useForm<FormInputs>({
     defaultValues,
   });
 
+  const previewValues = useWatch({ control });
+
+  const previewShortlink: ShortlinkType = {
+    id: selectedShortlink?.id ?? "preview",
+    subtitle: previewValues.subtitle ?? "",
+    title: previewValues.title ?? "Untitled...",
+    url: previewValues.url ?? "",
+    icon: previewValues.icon
+      ? previewValues.icon
+      : selectedShortlink
+        ? getFavicon(selectedShortlink.url)
+        : undefined,
+  };
+
   const onSubmit: SubmitHandler<FormInputs> = async (data) => {
-    const normalizedUrl = normalizeUrl(data.URL);
+    const normalizedUrl = normalizeUrl(data.url);
     reset(defaultValues);
 
-    await addMutation.mutateAsync({
+    const mutation = selectedShortlink ? editMutation : addMutation;
+
+    const payload: ShortlinkType = {
       id: selectedShortlink?.id ?? uuidv4(),
-      subtitle: String(data.Subtitle),
-      title: data.Title,
+      subtitle: String(data.subtitle),
+      title: data.title,
       url: normalizedUrl,
-    });
+      icon: data?.icon ? data.icon : undefined,
+    };
+
+    await mutation.mutateAsync(payload);
 
     onCloseModal();
   };
@@ -86,38 +128,63 @@ const ShortlinksModal = () => {
           : "Save a link for later. You know, so you might be able to actually find it when you need it."
       }
       onClose={onCloseModal}
+      className="max-w-3xl"
     >
-      <form className="mt-6 flex flex-col" onSubmit={handleSubmit(onSubmit)}>
-        <FormField
-          label="URL"
-          register={register}
-          required
-          pattern={
-            /^(https?:\/\/)?(www\.)?[a-zA-Z0-9-]+(\.[a-zA-Z]{2,})+(\/\S*)?$/
-          }
-        />
-        <FormField label="Title" register={register} required />
-        <FormField label="Subtitle" register={register} required={false} />
-
-        <div className="mt-6 flex justify-end gap-2">
-          {selectedShortlink && (
-            <Button
-              variant={buttonVariants.secondary}
-              label="Delete"
-              onClick={async (event) => {
-                event.preventDefault();
-                if (!selectedShortlink?.id) {
-                  return;
-                }
-
-                await deleteMutation.mutateAsync(selectedShortlink.id);
-                onCloseModal();
-              }}
-            />
-          )}
-          <Button label={selectedShortlink ? "Update link" : "Save link"} />
+      <div className="mt-4 grid grid-cols-5">
+        <div className="pointer-events-none col-span-2 flex flex-col gap-2 self-start rounded bg-gray-100 p-3">
+          <p className="text-sm font-semibold text-gray-500">Preview</p>
+          <Shortlink item={previewShortlink} isDragging={true} />
         </div>
-      </form>
+        <form
+          className="col-span-3 flex flex-col"
+          onSubmit={handleSubmit(onSubmit)}
+        >
+          <FormField
+            autoFocus
+            label="url"
+            register={register}
+            error={errors.url}
+            required
+            pattern={
+              /^(https?:\/\/)?(www\.)?[a-zA-Z0-9-]+(\.[a-zA-Z]{2,})+(\/\S*)?$/
+            }
+          />
+          <FormField
+            label="title"
+            register={register}
+            required
+            error={errors.title}
+          />
+          <FormField
+            label="subtitle"
+            register={register}
+            error={errors.subtitle}
+          />
+          <FormField label="icon" register={register} error={errors.icon} />
+
+          <div className="mt-6 flex justify-end gap-2">
+            {selectedShortlink && (
+              <Button
+                variant={buttonVariants.secondary}
+                label="Delete"
+                type="button"
+                onClick={async (event) => {
+                  event.preventDefault();
+                  event.stopPropagation();
+
+                  if (!selectedShortlink?.id) {
+                    return;
+                  }
+
+                  await deleteMutation.mutateAsync(selectedShortlink.id);
+                  onCloseModal();
+                }}
+              />
+            )}
+            <Button label={selectedShortlink ? "Update link" : "Save link"} />
+          </div>
+        </form>
+      </div>
     </Modal>
   );
 };