+++ src/components/Shortlinks/partials/Shortlink.tsx
@@ -3,6 +3,7 @@ import { EllipsisHorizontalIcon } from "@heroicons/react/24/solid";
 import { Modals, useModalStore } from "@/src/hooks/stores/useModalStore";
 import { useShortlinkStore } from "@/src/hooks/stores/useShortlinkStore";
 import { Shortlink } from "@/src/services/shortlinks";
+import { cn } from "@/src/utils";
 import { getFavicon } from "@/src/utils/getFavicon";
 
 import IconButton from "../../ui/IconButton";
@@ -10,19 +11,23 @@ import ImageWithFallback from "../../ui/ImageWithFallback";
 
 type Props = {
   item: Shortlink;
+  isDragging: boolean;
 };
 
-const Shortlink = ({ item }: Props) => {
+const Shortlink = ({ item, isDragging }: Props) => {
   const setActiveModal = useModalStore((state) => state.setActiveModal);
   const setSelectedShortlink = useShortlinkStore((state) => state.setSelected);
 
   return (
     <a
-      href={item.url}
-      className="flex cursor-pointer items-center gap-3 rounded border px-3 py-2.5 shadow ring-sky-500 duration-150 hover:ring focus:outline-none focus:ring"
+      href={isDragging ? undefined : item.url}
+      className={cn(
+        "w-full flex cursor-pointer items-center gap-3 rounded border bg-white px-3 py-2.5 shadow ring-sky-500 duration-150 hover:ring focus:outline-none focus:ring",
+        isDragging && "cursor-grabbing",
+      )}
     >
       <ImageWithFallback
-        src={getFavicon(item.url)}
+        src={item.icon ?? getFavicon(item.url)}
         fallbackSrc="/icon-34.png"
         className="size-6 shrink-0 rounded"
       />
@@ -31,10 +36,14 @@ const Shortlink = ({ item }: Props) => {
         <p className="truncate text-sm text-gray-600">{item.subtitle}</p>
       </div>
       <IconButton
-        className="border-none shadow-none"
+        className="border-none shadow-none "
         onClick={(event) => {
           event.preventDefault();
           event.stopPropagation();
+
+          if (isDragging) {
+            return;
+          }
 
           setSelectedShortlink(item);
           setActiveModal(Modals.shortlink);