+++ src/services/shortlinks/editShortlink.ts
@@ -5,9 +5,10 @@ import { Shortlink } from "./types";
 
 export const editShortlink = async (editedItem: Shortlink) => {
   const existingItems = await getShortlinks();
-  const updatedItems = [
-    ...existingItems.filter((item) => item.id !== editedItem.id),
-    editedItem,
-  ];
+
+  const updatedItems = existingItems.map((item) =>
+    item.id === editedItem.id ? editedItem : item
+  );
+
   await storage.local.set({ shortlinks: updatedItems });
 };