+++ src/services/shortlinks/index.ts
@@ -2,4 +2,5 @@ export * from "./addShortlink";
 export * from "./editShortlink";
 export * from "./deleteShortlink";
 export * from "./getShortlinks";
+export * from './reorderShortlinks';
 export * from "./types";